Correctly handle the child termination event
When a child terminates, we will have to start cleaning up, which means
that for the FirmwareFSM, shutdown all other children. Once the
FirmwareFSM has terminated, the ModemFirmwareActor will, which will then
trigger a s hutdhown of the ModemActor (and its ModemFSM), which then
will shut down the open serial port.